% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesVAR.R
\name{BayesVAR}
\alias{BayesVAR}
\title{Bayesian Estimation of VAR}
\usage{
BayesVAR(
  data,
  plag = 2,
  iter = 10000,
  burnin = 5000,
  prior = "none",
  type = "none"
)
}
\arguments{
\item{data}{a matrix which include all endogenous variables in VAR}

\item{plag}{a lag order in VAR}

\item{iter}{iterations of the MCMC}

\item{prior}{prior setting for VAR}

\item{type}{\code{'none','drift'} or \code{'trend'}}
}
\description{
Estimate a VAR base on bayesian method
}
