% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irfFAVAR.R
\name{irfFAVAR}
\alias{irfFAVAR}
\title{Impulse Response Function for FAVAR}
\usage{
irfFAVAR(fit, nvar = 116)
}
\arguments{
\item{fit}{FAVAR object.}

\item{nvar}{whose impulse response is computed? It's coloumn index in X.}
}
\value{
a object from \code{ggplot2::ggplot}.
}
\description{
Impulse Response Function for FAVAR
}
