% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slowcode.R
\docType{data}
\name{slowcode}
\alias{slowcode}
\title{Slow-moving or Not}
\format{
An object of class \code{logical} of length 115.
}
\source{
\url{https://sites.google.com/site/garykoop/home/computer-code-2}
}
\usage{
slowcode
}
\description{
A logic vector, record the variables that are the 1st cloumn through the 115th coloumn
in \code{regdata} is slow-moving or not.
}
\keyword{datasets}
