% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tcode.R
\docType{data}
\name{tcode}
\alias{tcode}
\title{Transformation Form for X}
\format{
An object of class \code{character} of length 118.
}
\source{
\url{https://sites.google.com/site/garykoop/home/computer-code-2}
}
\usage{
tcode
}
\description{
Record the transformation form for the 1st column through the 115th column
in \code{regdata}, and 'level' is Level, 'ln' is logarithm, 'Dln' is first difference of logarithm.
}
\keyword{datasets}
