% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BGM.R
\name{BGM}
\alias{BGM}
\title{Separate \code{R} From \code{X}}
\usage{
BGM(X, R, K = 2, niter = 10)
}
\arguments{
\item{X}{a large matrix from which principle components are extrated.}

\item{R}{a numerica vector which we are interesting in, for example interest rates.}

\item{K}{the number of extracted princple components.}

\item{niter}{the iterations, see details.}
}
\description{
Separate \code{R} From \code{X}
}
