% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef.favar.R
\name{coef.favar}
\alias{coef.favar}
\title{Extract Coefficients of a FAVAR Model}
\usage{
\method{coef}{favar}(object, ...)
}
\arguments{
\item{object}{a class 'favar'.}

\item{...}{additional arguments affecting the coefficients produced.}
}
\value{
A list\describe{
\item{fct_loading}{Factor loading matrix in a factor equation.}
\item{varcoef}{regression coefficients in VAR equations.}
}
}
\description{
Extract Coefficients of a FAVAR Model
}
