% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FAVAR.R
\name{FAVAR}
\alias{FAVAR}
\title{FAVAR}
\usage{
FAVAR(
  Y,
  X,
  slowcode,
  standerze = TRUE,
  fctmethod = "BBE",
  varprior = "none",
  nrep = 15000,
  nburn = 5000,
  K = 2,
  plag = 2,
  nhor = NULL,
  delta = 2.73
)
}
\arguments{
\item{slowcode}{a logical vector}

\item{standerze}{logical value, wheather standarze X and Y}

\item{fctmethod}{\code{'BBE'} or \code{'BGM'}. \code{'BBE'}(default) means the fators extracted method by Bernanke et al. (2005),
and \code{'BGM'} means the fators extracted method by Boivin et al. (2009).}

\item{K}{number of factors.}

\item{plag}{lag order in the VAR equation}

\item{nhor}{IRF horizon, default is \code{NULL}}

\item{\code{'none'}}{or \code{'mn'}, varprior prior setting for VAR. \code{'none'} means noninformative prior, and
\code{'mn'} means Minnesota prior.}
}
\description{
Estimate a FAVAR by Bernanke et al. (2005).
}
\references{
\link{1} Bernanke, B.S., J. Boivin and P. Eliasz, Measuring the Eeefects of Monetary Policy:
A Factor-Augmented Vector Autoregressive (FAVAR) Approach. Quarterly Journal of Economics, 2005. 120(1): p. 387-422.

\link{2} Boivin, J., M.P. Giannoni and I. Mihov, Sticky Prices and Monetary Policy: Evidence
from Disaggregated US Data. American Economic Review, 2009. 99(1): p. 350-384.
}
