% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/irfFAVAR.R
\name{irfFAVAR}
\alias{irfFAVAR}
\title{Impulse Response Function for FAVAR}
\usage{
irfFAVAR(fit, nvar = 116)
}
\value{
a object from \code{ggplot2::ggplot}.
}
\description{
Impulse Response Function for FAVAR
}
