% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryFAVAR.R
\name{summary.favar}
\alias{summary.favar}
\title{Print Results of FAVAR}
\usage{
\method{summary}{favar}(favar, xvar = NULL)
}
\arguments{
\item{favar}{a object from function \code{FAVAR}.}

\item{xvar}{a numeric vector, which variables in \code{X} was printed. It's a index.
If it's \code{NULL}, estimation results for X = F + Y is not printed.}
}
\description{
Print Results of FAVAR
}
